+++ src/Form/InputHidden.php
@@ -4,60 +4,34 @@ declare(strict_types=1);
 
 namespace UIAwesome\Html\Form;
 
-use UIAwesome\Html\Attribute\Form\{
-    HasAutocomplete,
-    HasForm,
-    HasList,
-    HasMaxlength,
-    HasMinlength,
-    HasPattern,
-    HasPlaceholder,
-    HasReadonly,
-    HasRequired,
-    HasSize
-};
-use UIAwesome\Html\Attribute\Global\{CanBeAutofocus, HasSpellcheck, HasTabindex};
+use UIAwesome\Html\Attribute\Form\{HasAutocomplete, HasForm};
 use UIAwesome\Html\Attribute\HasValue;
 use UIAwesome\Html\Attribute\Values\Type;
 use UIAwesome\Html\Core\Element\BaseInput;
 use UIAwesome\Html\Interop\{VoidInterface, Voids};
 
 /**
- * Renders the HTML `<input type="tel">` element for telephone number input.
+ * Renders the HTML `<input type="hidden">` element.
  *
- * The value uses the `tel` format (for example, `123-456-7890`).
+ * The `<input type="hidden">` defines a hidden input field.
  *
  * Usage example:
  * ```php
- * echo \UIAwesome\Html\Form\InputTel::tag()
- *     ->name('phone')
- *     ->placeholder('123-456-7890')
- *     ->pattern('[0-9]{3}-[0-9]{3}-[0-9]{4}')
- *     ->size(20)
- *     ->required(true)
+ * echo \UIAwesome\Html\Form\InputHidden::tag()
+ *     ->name('csrf_token')
+ *     ->value('1234567890')
  *     ->render();
  * ```
  *
- * @link https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/input/tel
+ * @link https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/input/hidden
  *
  * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
-final class InputTel extends BaseInput
+final class InputHidden extends BaseInput
 {
-    use CanBeAutofocus;
     use HasAutocomplete;
     use HasForm;
-    use HasList;
-    use HasMaxlength;
-    use HasMinlength;
-    use HasPattern;
-    use HasPlaceholder;
-    use HasReadonly;
-    use HasRequired;
-    use HasSize;
-    use HasSpellcheck;
-    use HasTabindex;
     use HasValue;
 
     /**
@@ -79,7 +53,7 @@ final class InputTel extends BaseInput
      */
     protected function loadDefault(): array
     {
-        return parent::loadDefault() + ['type' => [Type::TEL]];
+        return parent::loadDefault() + ['type' => [Type::HIDDEN]];
     }
 
     /**